requirements.txt
# AW-SafeSeg (IDDAW) - Complete Requirements
# Install with: pip install -r requirements.txt

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=9.5.0
scikit-image>=0.21.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Web Framework (Backend)
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0

# Authentication
PyJWT>=2.8.0
bcrypt>=4.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Optional: For better performance
# albumentations>=1.3.0  # Advanced augmentations
# tensorboard>=2.13.0    # Training visualization
